refactor: UserOrganizationService 인터페이스 분리 #26
